extracted out participant table from the participant list
